+++ flatpages_i18n/admin.py
@@ -1,83 +1,40 @@
-from builtins import str as text
-from django.conf import settings
 from django.contrib import admin
-from django.contrib.admin.sites import NotRegistered
-from django.db.models import TextField
 from django.utils.translation import ugettext_lazy as _
+from mptt.admin import DraggableMPTTAdmin
 
-from modeltranslation.admin import TranslationAdmin
-from mptt.admin import MPTTModelAdmin
-
-from flatpages_i18n.forms import FlatpageForm, MenuItemForm
-from flatpages_i18n.models import FlatPage_i18n, MenuItem
-from flatpages_i18n.widgets import RedactorEditor
+from flatpages_i18n.models import FlatPage_i18n, MenuItem, Menu
 
 
-class FlatPageAdmin(MPTTModelAdmin, TranslationAdmin):
-    form = FlatpageForm
-    mptt_level_indent = 0
-
+@admin.register(FlatPage_i18n)
+class FlatPageAdmin(admin.ModelAdmin):
     fieldsets = (
-        (None, {'fields': ('parent', 'machine_name', 'url', 'title', 'content', 'sites', )}),
+        (None, {'fields': ('title_i18n', 'slug_i18n', 'machine_name', 'content_i18n', 'sites', )}),
         (_(u'Advanced options'), {
             'classes': ('collapse',),
             'fields': ('registration_required', 'template_name')
         })
     )
-    list_display = ('indented_title', 'url', 'parent', 'weight')
+    list_display = ('title_i18n', 'slug_i18n', 'machine_name')
     list_filter = ('sites', 'registration_required')
-    list_editable = ['parent', 'weight']
-    search_fields = ('url', 'title')
+    search_fields = ('slug_i18n', 'title_i18n', 'content_i18n', 'machine_name')
     readonly_fields = ('created', 'modified')
 
-    class Media:
-        js = ('js/flatpages_i18n/jquery.js', )
-        css = {
-            'all': ('css/flatpages_i18n/admin.css', )
-        }
 
-    def __init__(self, *args, **kwargs):
-        if getattr(settings, 'FLATPAGES_EDITOR', None) == 'REDACTOR':
-            self.formfield_overrides = {
-                TextField: {'widget': RedactorEditor},
-            }
-        super(TranslationAdmin, self).__init__(*args, **kwargs)
-        self._patch_list_editable()
-
-    def indented_title(self, obj):
-        level = getattr(obj, obj._mptt_meta.level_attr)
-
-        if level is 0:
-            return obj
-
-        level_indicator = ''.join(['-' for i in range(level)])
-        return u'%s %s' % (level_indicator, text(obj))
-
-admin.site.register(FlatPage_i18n, FlatPageAdmin)
+@admin.register(Menu)
+class MenuAdmin(admin.ModelAdmin):
+    list_display = ('machine_name',)
+    list_display_links = ('machine_name',)
 
 
-class MenuItemAdmin(MPTTModelAdmin, TranslationAdmin):
-    form = MenuItemForm
-    mptt_level_indent = 0
+@admin.register(MenuItem)
+class MenuItemAdmin(DraggableMPTTAdmin):
+    search_fields = ('title_i18n', 'custom_link')
+    list_select_related = ['flatpage', 'menu']
+    list_display = ('tree_actions', 'indented_title', 'menu', 'title_i18n', 'custom_link_i18n', 'flatpage', 'url')
+    list_display_links = ('indented_title',)
+    list_editable = ['menu', 'flatpage', 'title_i18n', 'custom_link_i18n']
+    list_filter = ['menu']
+    fields = ['menu', 'title_i18n', 'flatpage', 'custom_link_i18n', 'target']
 
-    fields = ['parent', 'title', 'flatpage', 'machine_name', 'has_custom_link', 'custom_link', 'target']
-    list_display = ('indented_title', 'machine_name', 'parent', 'weight')
-    list_editable = ['parent', 'weight']
-
-    def indented_title(self, obj):
-        level = getattr(obj, obj._mptt_meta.level_attr)
-
-        if level is 0:
-            return obj
-
-        level_indicator = ''.join(['-' for i in range(level)])
-        return u'%s %s' % (level_indicator, text(obj))
-
-admin.site.register(MenuItem, MenuItemAdmin)
-
-
-try:
-    from django.contrib.flatpages.models import FlatPage
-    admin.site.unregister(FlatPage)
-except (ImportError, NotRegistered) as e:
-    pass
+    def url(self, obj):
+        return obj.get_absolute_url()