+++ flatpages_i18n/urls.py
@@ -1,16 +1,9 @@
-from django.conf.urls import url
+from django.urls import path
 
-from flatpages_i18n.views import redactor_upload
-from flatpages_i18n.forms import FileForm, ImageForm
+from flatpages_i18n.views import DetailView
 
-urlpatterns = (
-    url('^upload/image/(?P<upload_to>.*)', redactor_upload, {
-        'form_class': ImageForm,
-        'response': lambda name, url: '<img src="%s" alt="%s" />' % (url, name),
-        }, name='redactor_upload_image'),
+app_name = 'flatpages_i18n'
 
-    url('^upload/file/(?P<upload_to>.*)', redactor_upload, {
-        'form_class': FileForm,
-        'response': lambda name, url: '<a href="%s">%s</a>' % (url, name),
-        }, name='redactor_upload_file'),
-)
+urlpatterns = [
+    path('<str:slug>/', DetailView.as_view(), name='detail')
+]